fix: rename routes
