+++ apps/ui/src/router.ts
@@ -1,5 +1,5 @@
 import { createRouter, createWebHashHistory } from 'vue-router';
-import Home from '@/views/Home.vue';
+import Landing from '@/views/Landing.vue';
 import Space from '@/views/Space.vue';
 import SpaceOverview from '@/views/Space/Overview.vue';
 import SpaceProposals from '@/views/Space/Proposals.vue';
@@ -17,15 +17,15 @@ import Create from '@/views/Create.vue';
 import Settings from '@/views/Settings.vue';
 import Contacts from '@/views/Settings/Contacts.vue';
 import Explore from '@/views/Explore.vue';
+import Home from '@/views/Home.vue';
 import SettingsSpaces from '@/views/Settings/Spaces.vue';
-import SettingsFollowings from '@/views/Settings/Followings.vue';
 import Apps from '@/views/Apps.vue';
 import App from '@/views/App.vue';
 
 const { mixpanel } = useMixpanel();
 
 const routes: any[] = [
-  { path: '/', name: 'home', component: Home },
+  { path: '/', name: 'landing', component: Landing },
   {
     path: '/:id',
     name: 'space',
@@ -62,11 +62,11 @@ const routes: any[] = [
     component: Settings,
     children: [
       { path: '', name: 'settings-spaces', component: SettingsSpaces },
-      { path: 'home', name: 'settings-followings', component: SettingsFollowings },
       { path: 'contacts', name: 'settings-contacts', component: Contacts }
     ]
   },
   { path: '/explore', name: 'explore', component: Explore },
+  { path: '/home', name: 'home', component: Home },
   { path: '/apps', name: 'apps', component: Apps },
   { path: '/apps/:id', name: 'app', component: App }
 ];
